Correct solution to exercise 2.69
Use huf-adjoin-set instead of append to add merged tree to the set of
nodes that are left to merge. Using append always puts it at the end,
which doesn't always keep the set ordered and results in sub-optimal
trees being constructed. huf-adjoin-set inserts the new tree just before
any tree that has more weight, keeping the set ordered.

My solution was working in the sense that the constructed tree could be
used to encode and decode successfully, but the encoded message was
larger than it could have been.